backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.1
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment & Configuration
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI Model Dependencies
gradio-client==0.5.0
nltk==3.8.1
deep-translator==1.11.4
langdetect==1.0.9

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1